Add Logger to each service

+++ src/main/java/com/capstone/kelompok10/service/implementation/InstructorServiceImpl.java
@@ -9,6 +9,8 @@ import com.capstone.kelompok10.model.entity.InstructorEntity;
 import com.capstone.kelompok10.repository.InstructorRepository;
 import com.capstone.kelompok10.service.interfaces.InstructorService;
 
+import lombok.extern.log4j.Log4j2;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -16,6 +18,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
+@Log4j2
 public class InstructorServiceImpl implements InstructorService {
     InstructorRepository instructorRepository;
 
@@ -26,6 +29,7 @@ public class InstructorServiceImpl implements InstructorService {
     
     @Override
     public List<InstructorEntity> findAll() {
+        log.info("Get all Instructor without DTO");
         List<InstructorEntity> instructor = new ArrayList<>();
         instructorRepository.findAll().forEach(instructor::add);
         return instructor;
@@ -33,18 +37,21 @@ public class InstructorServiceImpl implements InstructorService {
     
     @Override
     public Page<InstructorEntity> findAllPagination(int offset, int pageSize) {
+        log.info("Get all Instructor with Pagination");
         Page<InstructorEntity> instructor = instructorRepository.findAll(PageRequest.of(offset, pageSize));
         return instructor;
     }
 
     @Override
     public Page<InstructorEntity> findAllPaginationSort(int offset, int pageSize, String field){
+        log.info("Get all Instructor with Pagination and Sorting");
         Page<InstructorEntity> instructor = instructorRepository.findAll(PageRequest.of(offset, pageSize).withSort(Sort.by(field)));
         return instructor;
     }
 
     @Override
     public List<InstructorDtoGet> findAllDto() {
+        log.info("Get all Instructor with DTO");
         List<InstructorEntity> instructors = instructorRepository.findAll();
         List<InstructorDtoGet> instructorDtos = new ArrayList<>();
 
@@ -62,31 +69,59 @@ public class InstructorServiceImpl implements InstructorService {
 
     @Override
     public InstructorEntity getInstructorById(Long instructorId) {
-        return instructorRepository.findById(instructorId).get();
+        if(instructorRepository.findById(instructorId) != null){
+            log.info("Instructor with id {} found", instructorId);
+            return instructorRepository.findById(instructorId).get();
+        }else{
+            log.info("Instructor with id {} not found", instructorId);
+            throw new IllegalStateException("Instructor not found");
+        }
     }
 
     @Override
     public void updateInstructor(Long instructorId, InstructorDtoPost instructorDtoPost) {
-        InstructorEntity instructor2 = instructorRepository.findById(instructorId).get();
-        instructor2.setName(instructorDtoPost.getName());
-        instructor2.setContact(instructorDtoPost.getContact());
-        instructor2.setImageUrl(instructorDtoPost.getImageUrl());
-        instructorRepository.save(instructor2);
+        if(instructorRepository.findById(instructorId) != null){
+            InstructorEntity instructor2 = instructorRepository.findById(instructorId).get();
+            instructor2.setName(instructorDtoPost.getName());
+            instructor2.setContact(instructorDtoPost.getContact());
+            instructor2.setImageUrl(instructorDtoPost.getImageUrl());
+            instructorRepository.save(instructor2);
+            log.info("Instructor updated");
+        }else{
+            log.info("Instructor with id {} not found", instructorId);
+            throw new IllegalStateException("Instructor not found");
+        }
     }
 
     @Override
     public void deleteInstructor(Long instructorId) {
-        instructorRepository.deleteById(instructorId);
+        if(instructorRepository.findById(instructorId) != null){
+            log.info("Instructor with id {} successfully deleted", instructorId);
+            instructorRepository.deleteById(instructorId);
+        }else{
+            log.info("Instructor with id {} not found", instructorId);
+            throw new IllegalStateException("Instructor not found");
+        }
     }
 
 	@Override
 	public void createInstructorDto(InstructorDtoPost instructorDtoPost) {
-		InstructorEntity instructorEntity = new InstructorEntity();
+        InstructorEntity instructorEntity = new InstructorEntity();
         instructorEntity.setName(instructorDtoPost.getName());
         instructorEntity.setContact(instructorDtoPost.getContact());
         instructorEntity.setImageUrl(instructorDtoPost.getImageUrl());
-        
-		instructorRepository.save(instructorEntity);
-	}
+            
+        instructorRepository.save(instructorEntity);
+        log.info("Instructor created");
+    }
+
+    @Override
+    public Boolean instructorExist(Long instructorId) {
+        if(instructorRepository.findById(instructorId) == null){
+            return false;
+        }else{
+            return true;
+        }
+    }
 }
