+++ src/main/java/com/capstone/kelompok10/service/implementation/MemberServiceImpl.java
@@ -16,7 +16,9 @@ import com.capstone.kelompok10.repository.MemberRepository;
 import com.capstone.kelompok10.service.interfaces.MemberService;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
+@Log4j2
 @Service
 @RequiredArgsConstructor
 public class MemberServiceImpl implements MemberService {
@@ -29,6 +31,7 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public List<MemberEntity> findAll() {
+        log.info("Get all Member without DTO");
         List<MemberEntity> member = new ArrayList<>();
         memberRepository.findAll().forEach(member::add);
         return member;
@@ -36,18 +39,21 @@ public class MemberServiceImpl implements MemberService {
     
     @Override
     public Page<MemberEntity> findAllPagination(int offset, int pageSize) {
+        log.info("Get all Member with Pagination");
         Page<MemberEntity> member = memberRepository.findAll(PageRequest.of(offset, pageSize));
         return member;
     }
 
     @Override
     public Page<MemberEntity> findAllPaginationSort(int offset, int pageSize, String field){
+        log.info("Get all Member with Pagination and Sorting");
         Page<MemberEntity> member = memberRepository.findAll(PageRequest.of(offset, pageSize).withSort(Sort.by(field)));
         return member;
     }
 
     @Override
     public List<MemberDtoGet> findAllDto() {
+        log.info("Get all Member with DTO");
         List<MemberEntity> members = memberRepository.findAll();
         List<MemberDtoGet> memberDtos = new ArrayList<>();
         
@@ -64,29 +70,62 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public MemberEntity getMemberById(Long memberId) {
-        return memberRepository.findById(memberId).get();
+        if(memberRepository.findById(memberId) != null){
+            log.info("Member with id {} found", memberId);
+            return memberRepository.findById(memberId).get();
+        }else{
+            log.info("Member with id {} not found", memberId);
+            throw new IllegalStateException("Member not found");
+        }
     }
 
     @Override
     public void updateMember(Long memberId, MemberDtoPost memberDtoPost) {
-        MemberEntity member2 = memberRepository.findById(memberId).get();
-        member2.setLength(memberDtoPost.getLength());
-        member2.setPrice(memberDtoPost.getPrice());
+        if(memberRepository.findById(memberId) != null){
+            MemberEntity member2 = memberRepository.findById(memberId).get();
+            member2.setLength(memberDtoPost.getLength());
+            member2.setPrice(memberDtoPost.getPrice());
 
-        memberRepository.save(member2);
+            memberRepository.save(member2);
+            log.info("Member updated");
+        }else{
+            log.info("Member with id {} not found", memberId);
+            throw new IllegalStateException("Member not found");
+        }
     }
 
     @Override
     public void deleteMember(Long memberId) {
-        memberRepository.deleteById(memberId);
+        if(memberRepository.findById(memberId) != null){
+            memberRepository.deleteById(memberId);
+            log.info("Member with id {} successfully deleted", memberId);
+        }else{
+            log.info("Member with id {} not found", memberId);
+            throw new IllegalStateException("Member not found");
+        }
     }
 
 	@Override
 	public void createMemberDto(MemberDtoPost memberDtoPost) {
-		MemberEntity memberEntity = new MemberEntity();
-        memberEntity.setLength(memberDtoPost.getLength());
-        memberEntity.setPrice(memberDtoPost.getPrice());
+        if(memberRepository.findByLength(memberDtoPost.getLength()) == null){
+            MemberEntity memberEntity = new MemberEntity();
+            memberEntity.setLength(memberDtoPost.getLength());
+            memberEntity.setPrice(memberDtoPost.getPrice());
 
-        memberRepository.save(memberEntity);
+            memberRepository.save(memberEntity);
+            log.info("Member created");
+        }else{
+            log.info("Member with length {} already exist");
+            throw new IllegalStateException("Member already exist");
+        }
 	}
+
+    @Override
+    public Boolean memberExist(Long memberId) {
+        if(memberRepository.findById(memberId) == null){
+            return false;
+        }else{
+            return true;
+        }
+    }
 }
