+++ src/main/java/com/capstone/kelompok10/service/implementation/MembershipServiceImpl.java
@@ -9,7 +9,11 @@ import com.capstone.kelompok10.model.entity.MemberEntity;
 import com.capstone.kelompok10.model.entity.MembershipEntity;
 import com.capstone.kelompok10.model.entity.UserEntity;
 import com.capstone.kelompok10.repository.MembershipRepository;
+import com.capstone.kelompok10.service.interfaces.MemberService;
 import com.capstone.kelompok10.service.interfaces.MembershipService;
+import com.capstone.kelompok10.service.interfaces.UserService;
+
+import lombok.extern.log4j.Log4j2;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -17,9 +21,16 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+@Log4j2
 @Service
 public class MembershipServiceImpl implements MembershipService {
     MembershipRepository membershipRepository;
+
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private MemberService memberService;
 
     @Autowired
     public MembershipServiceImpl(MembershipRepository membershipRepository){
@@ -28,6 +39,7 @@ public class MembershipServiceImpl implements MembershipService {
 
     @Override
     public List<MembershipEntity> findAll() {
+        log.info("Get all Membership without DTO");
         List<MembershipEntity> membership = new ArrayList<>();
         membershipRepository.findAll().forEach(membership::add);
         return membership;
@@ -35,18 +47,21 @@ public class MembershipServiceImpl implements MembershipService {
     
     @Override
     public Page<MembershipEntity> findAllPagination(int offset, int pageSize) {
+        log.info("Get all Membership with Pagination");
         Page<MembershipEntity> membership = membershipRepository.findAll(PageRequest.of(offset, pageSize));
         return membership;
     }
 
     @Override
     public Page<MembershipEntity> findAllPaginationSort(int offset, int pageSize, String field){
+        log.info("Get all Membership with Pagination and Sorting");
         Page<MembershipEntity> membership = membershipRepository.findAll(PageRequest.of(offset, pageSize).withSort(Sort.by(field)));
         return membership;
     }
 
     @Override
     public List<MembershipDtoGet> findAllDto() {
+        log.info("Get all Membership with DTO");
         List<MembershipEntity> memberships = membershipRepository.findAll();
         List<MembershipDtoGet> membershipDtos = new ArrayList<>();
         
@@ -64,29 +79,52 @@ public class MembershipServiceImpl implements MembershipService {
 
     @Override
     public MembershipEntity getMembershipById(Long membershipId) {
-        return membershipRepository.findById(membershipId).get();
+        if(membershipRepository.findById(membershipId) != null){
+            log.info("Membership with id {} found", membershipId);
+            return membershipRepository.findById(membershipId).get();
+        }else{
+            log.info("Membership with id {} not found", membershipId);
+            throw new IllegalStateException("Membership not Found");
+        }
     }
 
     @Override
     public void updateMembership(Long membershipId, MembershipDtoPost membershipDtoPost) {
-        MembershipEntity membership2 = membershipRepository.findById(membershipId).get();
+        if(membershipRepository.findById(membershipId) != null){
+            MembershipEntity membership2 = membershipRepository.findById(membershipId).get();
 
-        MemberEntity memberEntity = new MemberEntity();
-        memberEntity.setMemberId(membershipDtoPost.getMemberId());
+            MemberEntity memberEntity = new MemberEntity();
+            memberEntity.setMemberId(membershipDtoPost.getMemberId());
 
-        UserEntity userEntity = new UserEntity();
-        userEntity.setUserId(membershipDtoPost.getUserId()); 
-
-        membership2.setStatus(membershipDtoPost.getStatus());
-        membership2.setUser(userEntity);
-        membership2.setMember(memberEntity);
-
-        membershipRepository.save(membership2);
+            UserEntity userEntity = new UserEntity();
+            userEntity.setUserId(membershipDtoPost.getUserId());
+            
+            if(userService.userExist(membershipDtoPost.getUserId()) == true && memberService.memberExist(membershipDtoPost.getMemberId()) == true){
+                membership2.setStatus(membershipDtoPost.getStatus());
+                membership2.setUser(userEntity);
+                membership2.setMember(memberEntity);
+    
+                membershipRepository.save(membership2);
+                log.info("membership updated");
+            }else{
+                log.info("failed to update membership");
+                throw new IllegalStateException("User or Member did'ny exist");
+            }       
+        }else{
+            log.info("Membership with id {} not found", membershipId);
+            throw new IllegalStateException("Membership not Found");
+        }
     }
 
     @Override
     public void deleteMembership(Long membershipId) {
-        membershipRepository.deleteById(membershipId);
+        if(membershipRepository.findById(membershipId) != null){
+            log.info("Membership with id {} successfully deleted", membershipId);
+            membershipRepository.deleteById(membershipId);
+        }else{
+            log.info("Membership with id {} not found", membershipId);
+            throw new IllegalStateException("Membership not Found");
+        }
         
     }
 
@@ -100,10 +138,16 @@ public class MembershipServiceImpl implements MembershipService {
         MemberEntity memberEntity = new MemberEntity();
         memberEntity.setMemberId(membershipDtoPost.getMemberId());
 
-        membershipEntity.setStatus(membershipDtoPost.getStatus());
-        membershipEntity.setUser(userEntity);
-        membershipEntity.setMember(memberEntity);
+        if(userService.userExist(membershipDtoPost.getUserId()) == true && memberService.memberExist(membershipDtoPost.getMemberId()) == true){
+            membershipEntity.setStatus(membershipDtoPost.getStatus());
+            membershipEntity.setUser(userEntity);
+            membershipEntity.setMember(memberEntity);
 
-        membershipRepository.save(membershipEntity);
+            membershipRepository.save(membershipEntity);
+            log.info("membership created");
+        }else{
+            log.info("failed to create membership");
+            throw new IllegalStateException("User or Member did'ny exist");
+        }
     }
 }
