+++ src/main/java/com/capstone/kelompok10/service/implementation/RoleServiceImpl.java
@@ -17,7 +17,9 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
+@Log4j2
 @Service
 @RequiredArgsConstructor
 public class RoleServiceImpl implements RoleService {
@@ -33,6 +35,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public List<RoleEntity> findAll() {
+        log.info("Get all Role without DTO");
         List<RoleEntity> role = new ArrayList<>();
         roleRepository.findAll().forEach(role::add);
         return role;
@@ -40,18 +43,21 @@ public class RoleServiceImpl implements RoleService {
     
     @Override
     public Page<RoleEntity> findAllPagination(int offset, int pageSize) {
+        log.info("Get all Role with Pagination");
         Page<RoleEntity> role = roleRepository.findAll(PageRequest.of(offset, pageSize));
         return role;
     }
 
     @Override
     public Page<RoleEntity> findAllPaginationSort(int offset, int pageSize, String field){
+        log.info("Get all Role with Pagination and Sorting");
         Page<RoleEntity> role = roleRepository.findAll(PageRequest.of(offset, pageSize).withSort(Sort.by(field)));
         return role;
     }
 
     @Override
     public List<RoleDtoGet> findAllDto() {
+        log.info("Get all Role with DTO");
         List<RoleEntity> roles = roleRepository.findAll();
         List<RoleDtoGet> roleDtos = new ArrayList<>();
         
@@ -67,32 +73,57 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleEntity getRoleById(Long roleId) {
-        return roleRepository.findById(roleId).get();
+        if(roleRepository.findById(roleId) != null){
+            log.info("Role with id {} found", roleId);
+            return roleRepository.findById(roleId).get();
+        }else{
+            log.info("Role with id {} not Found");
+            throw new IllegalStateException("Role not Found");
+        }
     }
 
     @Override
     public void createRole(RoleEntity role) {
         roleRepository.save(role);
+        log.info("Role created");
     }
 
     @Override
     public void updateRole(Long roleId, RoleDtoPost roleDtoPost) {
-        RoleEntity role2 = roleRepository.findById(roleId).get();
-        role2.setName(roleDtoPost.getName());
-        roleRepository.save(role2);
+        if(roleRepository.findById(roleId) != null){
+            RoleEntity role2 = roleRepository.findById(roleId).get();
+            role2.setName(roleDtoPost.getName());
+            roleRepository.save(role2);
+            log.info("Role updated");
+        }else{
+            log.info("Role with id {} not Found");
+            throw new IllegalStateException("Role not Found");
+        }
     }
 
     @Override
     public void deleteRole(Long roleId) {
-        roleRepository.deleteById(roleId);  
+        if(roleRepository.findById(roleId) != null){
+            log.info("Role with id {} successfully deleted", roleId);
+            roleRepository.deleteById(roleId);  
+        }else{
+            log.info("Role with id {} not Found");
+            throw new IllegalStateException("Role not Found");
+        }
     }
 
 	@Override
 	public void createRoleDto(RoleDtoPost roleDtoPost) {
-		RoleEntity roleEntity = new RoleEntity();
-        roleEntity.setName(roleDtoPost.getName());
+        if(roleRepository.findByName(roleDtoPost.getName()) != null){
+            log.info("Role with name {} already exist", roleDtoPost.getName());
+            throw new IllegalStateException("Role already exist");
+        }else{
+            RoleEntity roleEntity = new RoleEntity();
+            roleEntity.setName(roleDtoPost.getName());
 
-		roleRepository.save(roleEntity);
+            roleRepository.save(roleEntity);
+            log.info("Role created");
+        }
 	}
 }
